refactor: moved misc classes to utility package
